% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend_output.R
\name{amend_output}
\alias{amend_output}
\title{amend_output}
\usage{
amend_output(output, fit, f_equ, temp, rate, try_test, augment, return_fit)
}
\arguments{
\item{output}{broom output of fit}

\item{fit}{the model output of the fitting process}

\item{f_equ}{equation with fitted parameters}

\item{temp}{temperature values of measurements}

\item{rate}{rate that changes with temperature}

\item{try_test}{did the model fitting succeed or produce an error?}

\item{augment}{add columns to the original dataset such as predictions, residuals and cluster assignments using package::broom (T/F)?}

\item{return_fit}{return the model object (T/F)?}
}
\value{
a data frame of, depending on augment argument, if FALSE, parameters, if TRUE, data with predicted values
}
\description{
Helper function that add terms to the broom output of fit
}
