% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend_output_tempfit.R
\name{amend_output_tempfit}
\alias{amend_output_tempfit}
\title{amend_output_tempfit}
\usage{
amend_output_tempfit(fit, temp, rate, try_test = NULL, output_type)
}
\arguments{
\item{fit}{the model output of the fitting process}

\item{temp}{temperature values of measurements}

\item{rate}{rate that changes with temperature}

\item{try_test}{did the model fitting succeed or produce an error?}

\item{output_type}{type of output to return}
}
\value{
a data frame of, depending on augment argument, if FALSE, parameters, if TRUE, data with predicted values
}
\description{
Helper function that add terms to the broom output of fit
}
