% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emiliania_huxleyi.R
\docType{data}
\name{Emiliania_huxleyi}
\alias{Emiliania_huxleyi}
\title{Temperature response of the growth rate of Emiliania_huxleyi}
\format{A data frame with 39 rows and 3 variables:
\describe{
  \item{temp}{temperature}
  \item{rate}{growth rate}
  ...
}}
\source{
\url{to_be_added}
}
\usage{
Emiliania_huxleyi
}
\description{
A data set containing the temperature response of the growth rate of Emiliania_huxleyi
}
\keyword{datasets}
