% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equ10.R
\name{equ10}
\alias{equ10}
\title{Equation 10}
\usage{
equ10(temp, rate, augment = F, return_fit = F)
}
\arguments{
\item{temp}{temperature (in Celsius)}

\item{rate}{rate measurement}

\item{augment}{logical wether the dataset with fits should be returned instead of the parameter values}

\item{return_fit}{logical wether the model fit object should be returned}
}
\value{
a data frame of, depending on augment argument, if FALSE, parameters, if TRUE, data with predicted values
}
\description{
Equation from Thomas et al. (2014)
}
