% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equ11.R
\name{equ11}
\alias{equ11}
\title{Equation 11}
\usage{
equ11(temp, rate, augment = F, return_fit = F)
}
\arguments{
\item{temp}{temperature (in Celsius)}

\item{rate}{rate measurement}

\item{augment}{logical wether the dataset with fits should be returned instead of the parameter values}

\item{return_fit}{logical wether the model fit object should be returned}
}
\value{
depends on augment: if false, fitting parameters or fitted data
}
\description{
Equation in Montagnes et al. 2008
}
\examples{
output <- with(Emiliania_huxleyi, equ11(temp=temp, rate=rate))
}
