% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equ13.R
\name{equ13}
\alias{equ13}
\title{Equation 13}
\usage{
equ13(temp, rate, augment = F, return_fit = F)
}
\arguments{
\item{temp}{temperature (in Celsius)}

\item{rate}{rate measurement}

\item{augment}{logical wether the dataset with fits should be returned instead of the parameter values}

\item{return_fit}{logical wether the model fit object should be returned}
}
\value{
depends on augment: if false, fitting parameters or fitted data
}
\description{
Equation in Ratkowsky et al. (1983)
}
\examples{
output <- with(Emiliania_huxleyi, equ14(temp=temp, rate=rate))
}
