% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equ4.R
\name{equ4}
\alias{equ4}
\title{Equation 4}
\usage{
equ4(temp, rate, augment = F, return_fit = F)
}
\arguments{
\item{temp}{temperature (in Celsius)}

\item{rate}{rate measurement}

\item{augment}{logical wether the dataset with fits should be returned instead of the parameter values}

\item{return_fit}{logical wether the model fit object should be returned}
}
\value{
a data frame of, depending on augment argument, if FALSE, parameters, if TRUE, data with predicted values
}
\description{
Equation 4 is model H in  Li & Dickie (1987) citing Hinshelwood (1947)
}
\examples{
output <- with(Emiliania_huxleyi, equ4(temp=temp, rate=rate))
}
