% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equ5.R
\name{equ5}
\alias{equ5}
\title{Equation 5}
\usage{
equ5(temp, rate, augment = F, return_fit = F)
}
\arguments{
\item{temp}{temperature (in Celsius)}

\item{rate}{rate measurement}

\item{augment}{logical wether the dataset with fits should be returned instead of the parameter values}

\item{return_fit}{logical wether the model fit object should be returned}
}
\value{
a data frame of, depending on augment argument, if FALSE, parameters, if TRUE, data with predicted values
}
\description{
Equation 5 is model J from Li & Dickie (1987) citing Johnson et al. (1942)  Does not currently work
}
\examples{
output <- with(Emiliania_huxleyi, equ5(temp=temp, rate=rate))
}
