% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equ6.R
\name{equ6}
\alias{equ6}
\title{Equation 6}
\usage{
equ6(temp, rate, augment = F, return_fit = F)
}
\arguments{
\item{temp}{temperature (in Celsius)}

\item{rate}{rate measurement}

\item{augment}{logical wether the dataset with fits should be returned instead of the parameter values}

\item{return_fit}{logical wether the model fit object should be returned}
}
\value{
a data frame of, depending on augment argument, if FALSE, parameters, if TRUE, data with predicted values
}
\description{
Equation 6
}
\examples{
output <- with(Emiliania_huxleyi, equ6(temp=temp, rate=rate))

}
