% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equ7.R
\name{equ7}
\alias{equ7}
\title{Equation 7}
\usage{
equ7(temp, rate, augment = F, return_fit = F)
}
\arguments{
\item{temp}{temperature (in Celsius)}

\item{rate}{rate measurement}

\item{augment}{logical wether the dataset with fits should be returned instead of the parameter values}

\item{return_fit}{logical wether the model fit object should be returned}
}
\value{
a data frame of, depending on augment argument, if FALSE, parameters, if TRUE, data with predicted values
}
\description{
Equation 7 from Montagnes et al (2008) citing Schoolfield et al. (1981)
}
\examples{
output <- with(Emiliania_huxleyi, equ7(temp=temp, rate=rate))
}
