% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempfit.R
\name{tempfit}
\alias{tempfit}
\title{tempfit}
\usage{
tempfit(temp, rate, equation = 10, output_type = "parameters", ...)
}
\arguments{
\item{temp}{temperature (in Celsius)}

\item{rate}{rate measurement}

\item{equation}{numeric value of equation}

\item{output_type}{one of "parameters", "augment","model_object"}

\item{...}{supplemental parameters for  minpack.lm::nlsLM (eg. control=minpack.lm::nls.lm.control(maxiter = 10^5)}
}
\value{
a data frame of, depending on augment argument, if FALSE, parameters, if TRUE, data with predicted values
}
\description{
Fits equations to temperature response data
}
